// frontend/src/client/types.gen.ts
export type IntakeDocumentPublic = {
    id: string;
    template_id: string;
    form_name: string;
    category: string;
    filename: string;
    status: string;
    packet_id: string;
    created_at?: (string | null);
};

export type IntakePacketCreate = {
    patient_id_number: string;
    patient_name: string;
    facility: string;
    admission_date: string;
    counselor_name: string;
    template_ids: Array<(string)>;
};

export type IntakePacketPublic = {
    patient_id_number: string;
    patient_name: string;
    facility: string;
    admission_date: string;
    counselor_name: string;
    id: string;
    status: string;
    owner_id: string;
    created_at?: (string | null);
    documents?: Array<IntakeDocumentPublic>;
};